requirements.txt
Django==3.9.6
requests
bs4
django-crispy-forms
djangorestframework
markdown
django-cors-headers